requirements.txt
# OptiRoute AI - Python Dependencies

# Core Framework
streamlit>=1.28.0

# LangChain and AI Providers
langchain>=0.1.0
langchain-openai>=0.0.2
langchain-groq>=0.0.1
langchain-core>=0.1.0

# Environment Management
python-dotenv>=1.0.0

# API and HTTP
openai>=1.6.0
httpx>=0.25.0

# Utilities
typing-extensions>=4.8.0
pydantic>=2.5.0
